requirements.txt
# Core dependencies
# Versions aligned with pyproject.toml
openai>=2.15.0
anthropic>=0.76.0
google-generativeai>=0.3.1
requests>=2.31.0
pydantic>=2.5.2
python-dotenv>=1.0.0

# Backend server
fastapi>=0.109.2
uvicorn>=0.27.0
websockets>=12.0

# Optional development dependencies
# pytest>=7.4.0
# black>=23.7.0
# isort>=7.0.0
# mypy>=1.5.1
